Demote the Surface enum class to enum
GCC was producing warnings because all of the stuff in the standard about how
enums can be packed into bitfields so long as there are enough bits provided to
store all of the values in the enum gets thrown out the window for enum class.
My only guess is because the enum class is implemented less transparently be an
underlying primitive type, which can even be specified by the programmer (e.g.
enum class foo: char{...};). The standards committee must have decided that it
was too scary to be allowed to force that into a smaller location. Seriously
though, standards committee; weak.

+++ src/mocc-core/ray.cpp
@@ -83,7 +83,7 @@ namespace mocc {
             cm_surf_fw_ = mesh.coarse_surf( cm_cell_fw_, s[0] );
             for( auto nseg: cm_nseg ) {
                 ns = mesh.coarse_norm_point( *p, octant, s );
-                for( int i=0; i<ns; i++ ) {
+                for( size_t i=0; i<ns; i++ ) {
                     surfs_fw.push_back(s[i]);
                 }
                 nsegs_fw.push_back( nseg );
@@ -108,7 +108,7 @@ namespace mocc {
             cm_surf_bw_ = mesh.coarse_surf( cm_cell_bw_, s[0] );
             for( auto nseg: cm_nseg ) {
                 ns = mesh.coarse_norm_point( *p, octant, s);
-                for( int i=0; i<ns; i++ ) {
+                for( size_t i=0; i<ns; i++ ) {
                     surfs_bw.push_back(s[i]);
                 }
                 nsegs_bw.push_back( nseg );
